파일 정리

+++ 공부/LV_0/크기가 작은 부분문자열.cpp
@@ -8,8 +8,8 @@ int solution(string t, string p) {
     int answer = 0;
     for (int i = 0; i <= t.size()-p.size(); ++i)
     {
-        int num = stoi(t.substr(i, p.size()));
-        if (num <= stoi(p))
+        long long  num = stol(t.substr(i, p.size()));
+        if (num <= stol(p))
         {
             answer++;
         }
